OBJECTIVE 25: Fix unused variables/parameters (66 -> 58 warnings)
Fixed 8 warnings by adding (void) casts:
- cllm_optimizer_wrapper.c: gradient_scale
- cllm_production.c: scale
- cllm_training_threaded.c: ff_hidden, ln, std
- cllm_training.c: layer, token_ids parameters
- tools/cllm_inference.c: sign comparison cast

Result: 13 warnings eliminated total (71 -> 58), 18% reduction

+++ tools/cllm_inference.c
@@ -109,7 +109,7 @@ static void generate_text_proper(CLLMInference* inference, const char* prompt,
             // Convert BigFixed** to usable form for sampling
             double max_prob = big_fixed_to_double(inference->logits[0]);
             next_token = 0;
-            for (int i = 1; i < model->vocab_size; i++) {
+            for (int i = 1; i < (int)model->vocab_size; i++) {
                 double prob = big_fixed_to_double(inference->logits[i]);
                 if (prob > max_prob) {
                     max_prob = prob;
